backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
groq==0.4.1
elevenlabs==0.2.26
python-multipart==0.0.6
httpx==0.25.2